requirements.txt
livekit-agents
livekit-plugins-google
livekit-plugins-noise-cancellation
livekit-plugins-silero
duckduckgo-search
langchain-community
requests
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib